Working form using controlled component input
Input is linked to state.input, with state.input being the source of
truth. Now need to handle the submit, which we are currently just
console.logging. AKA need to implement Overview component, which should
render all the tasks...

All tasks should could be kept internally as a state in Overview, but
I think it makes more sense that App.js as a whole knows about allTasks.
So App.js should probably keep track of this, then pass it as a prop to
Overview.js. Big Ups