Fixed List items not rendering sometimes and id
I was passing <li key={uuid()}>. But instead needed to pass
it as an arrow function key={() => uuid()}. The basic idea is that
the list didn't know when to render items sometimes... aka problem
with unique key.

I remember reading that props could never update. This is not a prop
since li's are not components, but I imagine its the same problem..
So basically the function was grabbing a uniqueId only once, and solved
using arrow function, so the arrow function always runs.

ALSO i moved the id generation into App, and inside App, I keep track of
new id generation for each newTask. Then I just pass it down as a prop